use unsigned for sizes

+++ src/buffer_wrapper.cpp
@@ -48,7 +48,7 @@ buffer_wrapper::operator bool() const
     return m_pimpl ? true : false;
 }
 
-size_t buffer_wrapper::get_size()
+unsigned buffer_wrapper::get_size()
 {
     return m_pimpl ? m_pimpl->get_size() : 0;
 }
