+++ src/buffer_wrapper_impl.h
@@ -12,7 +12,7 @@ struct buffer_wrapper::pimpl
 
     pimpl() = default;
 
-    pimpl(Device device, size_t size, buffer_type flags, bool isDest)
+    pimpl(Device device, unsigned size, buffer_type flags, bool isDest)
         : m_device(device)
         , m_size(size)
         , m_usage(getBufferUsageFromType(flags, isDest))
@@ -65,7 +65,7 @@ struct buffer_wrapper::pimpl
         return m_done;
     }
 
-    size_t get_size()
+    unsigned get_size()
     {
         return m_buffer.GetSize();
     }
